move all 1p basis components to new interface
