+++ src/b1pcomponents/README.md
@@ -1,0 +1,2 @@
+
+This folder contains various 1p basis component wrappers.
